third_party/WebKit/LayoutTests/pointer-lock/mouse-event-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
    var mouseEvent = document.createEvent("MouseEvent");
    shouldBeDefined("mouseEvent.webkitMovementX");
    shouldBeDefined("mouseEvent.webkitMovementY");
</script>
</body>
</html>
